refactor: hookにだせていない処理をリファクタ
